move Game updates to GameUpdate.ts
