+++ src/client/graphics/layers/TerritoryLayer.ts
@@ -1,12 +1,13 @@
 import { PriorityQueue } from "@datastructures-js/priority-queue";
-import { Cell, Game, Player, PlayerType, Unit, UnitType, UnitUpdate } from "../../../core/game/Game";
+import { Cell, Game, Player, PlayerType, Unit, UnitType } from "../../../core/game/Game";
+import { UnitUpdate } from "../../../core/game/GameUpdates";
 import { PseudoRandom } from "../../../core/PseudoRandom";
 import { colord, Colord } from "colord";
 import { Theme } from "../../../core/configuration/Config";
 import { Layer } from "./Layer";
 import { EventBus } from "../../../core/EventBus";
 import { AlternateViewEvent, DragEvent, MouseDownEvent } from "../../InputHandler";
-import { GameView, PlayerView } from "../../../core/GameView";
+import { GameView, PlayerView } from "../../../core/game/GameView";
 import { euclDistFN, TileRef } from "../../../core/game/GameMap";
 
 export class TerritoryLayer implements Layer {
