--- src/core/GameViewData.ts
@@ -1,4 +1,0 @@
-
-
-
-
